requirements.txt
click
gin-config
tensorflow-gpu